+++ OnlineStore/src/cliente.java
@@ -1,0 +1,2 @@
+public class cliente {
+}
